Add files via upload
This program prompts the user to enter in a word or phrase. Once the user has entered their word or phrase they are then prompted to type in a word or character they would like to search for in the phrase that they just entered. If the phrase contains what they are searching for the code will return "True," if it does not, it will return "False."